CFP2-178: Move user.model to a better location
